+++ src/textual/css/parse.py
@@ -4,8 +4,6 @@ from functools import lru_cache
 from pathlib import PurePath
 from typing import Iterator, Iterable, NoReturn
 
-from rich import print
-
 from .errors import UnresolvedVariableError
 from .types import Specificity3
 from ._styles_builder import StylesBuilder, DeclarationError
@@ -36,7 +34,6 @@ SELECTOR_MAP: dict[str, tuple[SelectorType, Specificity3]] = {
 
 @lru_cache(maxsize=1024)
 def parse_selectors(css_selectors: str) -> tuple[SelectorSet, ...]:
-
     if not css_selectors.strip():
         return ()