+++ src/textual/devtools/renderables.py
@@ -5,7 +5,6 @@ from pathlib import Path
 from typing import Iterable
 
 from importlib_metadata import version
-
 from rich.align import Align
 from rich.console import Console, ConsoleOptions, RenderResult
 from rich.markup import escape
@@ -15,8 +14,9 @@ from rich.style import Style
 from rich.styled import Styled
 from rich.table import Table
 from rich.text import Text
+from typing_extensions import Literal
+
 from textual._log import LogGroup
-from textual._typing import Literal
 
 DevConsoleMessageLevel = Literal["info", "warning", "error"]